pyproject.toml
[tool.poetry]
name = "sustentare"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
mysql = "^0.0.3"
pandas = "^2.2.0"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"